fixed 2 sizing bugs
First bug was an index error when attempting to resize on the player
count selection screen. This is because the resize_adjustment method was
attempting to reference buttons that hadn't yet spawned.

Second bug resulted from fixing the first: Adjusting the window on the
player # selection caused the game board screen to be incorrectly sized
until manually resized. Fixed this by calling  resize_adjustment after
destroying the widgets on the initial screen.